+++ lib/Screens/Login/user_login.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:linguotech/Screens/Translation_Screen/translate.dart';
 import 'package:linguotech/services/google_sign_in.dart';
+import 'package:linguotech/widgets/Translate/urdu_Text_Card.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class SignInScreen extends StatefulWidget {
