+++ lib/Screens/Summarization/summary.dart
@@ -1,4 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:linguotech/widgets/icons.dart';
+import 'package:linguotech/widgets/language_selectot.dart';
+
+void main() {
+  runApp(SummaryGenerator());
+}
 
 class SummaryGenerator extends StatelessWidget {
   @override
@@ -28,10 +34,11 @@ class _SummaryGeneratorScreenState extends State<SummaryGeneratorScreen> {
     return Scaffold(
       appBar: AppBar(
         title: Center(
-            child: Text(
-          'Text Summarization',
-          style: TextStyle(fontWeight: FontWeight.bold),
-        )),
+          child: Text(
+            'Text Summarization',
+            style: TextStyle(fontWeight: FontWeight.bold),
+          ),
+        ),
         backgroundColor: Colors.white,
       ),
       body: SingleChildScrollView(
@@ -40,107 +47,10 @@ class _SummaryGeneratorScreenState extends State<SummaryGeneratorScreen> {
           child: Container(
             child: Column(
               children: [
-                Column(
-                  children: [
-                    Container(
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          GestureDetector(
-                            onTap: () {
-                              changeLanguage('Urdu');
-                            },
-                            child: Container(
-                              padding: EdgeInsets.symmetric(
-                                  horizontal: 15, vertical: 8),
-                              decoration: BoxDecoration(
-                                border: selectedLanguage == 'Urdu'
-                                    ? Border(
-                                        bottom: BorderSide(
-                                          color: Colors
-                                              .black, // Adjust the underline color
-                                          width:
-                                              2.0, // Adjust the thickness of the underline
-                                        ),
-                                      )
-                                    : null, // No border when not selected
-                              ),
-                              child: Text(
-                                'Urdu',
-                                style: TextStyle(
-                                  color: selectedLanguage == 'Urdu'
-                                      ? Colors.orange
-                                      : Colors.black,
-                                  fontWeight: FontWeight.bold,
-                                  decoration: selectedLanguage == 'Urdu'
-                                      ? TextDecoration.underline
-                                      : TextDecoration.none,
-                                ),
-                              ),
-                            ),
-                          ),
-                          SizedBox(width: 20),
-                          Icon(
-                            Icons.compare_arrows, // Double-headed arrow icon
-                            size: 30,
-                            color: Colors.black,
-                          ),
-                          SizedBox(width: 20),
-                          GestureDetector(
-                            onTap: () {
-                              changeLanguage('English');
-                            },
-                            child: Container(
-                              padding: EdgeInsets.symmetric(
-                                  horizontal: 15, vertical: 8),
-                              decoration: BoxDecoration(
-                                border: selectedLanguage == 'English'
-                                    ? Border(
-                                        bottom: BorderSide(
-                                          color: Colors
-                                              .black, // Adjust the underline color
-                                          width:
-                                              2.0, // Adjust the thickness of the underline
-                                        ),
-                                      )
-                                    : null, // No border when not selected
-                              ),
-                              child: Text(
-                                'English',
-                                style: TextStyle(
-                                  color: selectedLanguage == 'English'
-                                      ? Colors.orange
-                                      : Colors.black,
-                                  fontWeight: FontWeight.bold,
-                                  decoration: selectedLanguage == 'English'
-                                      ? TextDecoration.underline
-                                      : TextDecoration.none,
-                                ),
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-                SizedBox(
-                  height: 10,
-                ),
-                Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Icon(Icons.link),
-                    SizedBox(width: 5),
-                    Text(
-                      'Add link here',
-                      style: TextStyle(fontWeight: FontWeight.bold),
-                    ),
-                  ],
-                ),
-                SizedBox(
-                  height: 20,
-                ),
+                LanguageSelector(selectedLanguage, changeLanguage),
+                SizedBox(height: 10),
+                LinkIcon(),
+                SizedBox(height: 20),
                 TextField(
                   controller: _inputController,
                   maxLines: 10,
